Added context menu to treeview.
